Minor updates and building out new feature
Added parsing from source port ONLY if it's using an operator and port in the ACE. example; eq www OR eq snmp OR gt 79
output.txt will not contain the source port column and currently no outputs utilize the source port.
ACE comparing is also not setup to use the source port yet. Still assinging value of 100 (assumed match of not used)
It IS, however in the debugdump.txt
To identify source ports using object-groups I will need the AccessListSubRule class to read the listOfObjects list when it breaks down the ACE. Technically we have that before the ACE's get parsed. Just have to figure some of the details. Update will be coming for that later.

Added incomplete function for 'localpoc'. incomplete but does not affect other functions. Also not listed in the menu. On a deadline.